+++ fdb-record-layer-core/src/main/java/com/apple/foundationdb/record/provider/foundationdb/FDBTransactionContext.java
@@ -73,6 +73,11 @@ public class FDBTransactionContext {
         return executor;
     }
 
+    @Nonnull
+    public CompletableFuture<Long> getApproximateTransactionSize() {
+        return transaction.getApproximateSize();
+    }
+
     @Nullable
     public FDBStoreTimer getTimer() {
         return timer;